Refactor UpdateResolver -> UpdateService
Added GetChangelog Method
